reorg: reduce duplication
There were three copies of all the music. This reduces it to just one copy,
and the music never contains a \paper call, so that can be handled generally.

More reorg is coming, as I prefer lyrics, music and structure to be separate
files so they can be independently changed. This allows for separate files for
midi and book while entirely reusing both the music and lyrics, as well as
causing songs with a common tune to use the same tune source file.

There are a lot of Lilypond warnings with the files as they stand, and that's
also going to be resolved.