bug: UTF-8 names don't play well with /include
We fixed some additional problems with the songs, but the big one was some renaming.

German words with umlauts can be converted without damaging the meaning (addition of "e")
but I didn't have a solution to Spanish accents, so (for not, at least) I've just dropped
the accent. Not preferred, but it makes it pass.

I've started picking the songs to think about for the next phase, but the current files
suffer from having no '/bookpart's, so they run together poorly.